+++ app/src/main/java/woowacourse/shopping/ui/main/ProductViewHolder.kt
@@ -1,4 +1,4 @@
-package woowacourse.shopping.ui
+package woowacourse.shopping.ui.main
 
 import android.view.LayoutInflater
 import android.view.ViewGroup
