+++ app/src/test/java/woowacourse/shopping/MainActivityTest.kt
@@ -8,8 +8,8 @@ import org.junit.Test
 import org.junit.runner.RunWith
 import org.robolectric.Robolectric
 import org.robolectric.RobolectricTestRunner
-import woowacourse.shopping.ui.MainActivity
-import woowacourse.shopping.ui.MainViewModel
+import woowacourse.shopping.ui.main.MainActivity
+import woowacourse.shopping.ui.main.MainViewModel
 
 
 @RunWith(RobolectricTestRunner::class)
